requirements.txt
Flask
mysql-connector-python
Flask-CORS
python-dotenv  # NOVO: Para carregar o .env
flask-bcrypt   # NOVO: Para criptografia de senhas
PyJWT          # NOVO: Para tokens de autenticação